refactor: product 엔티티 분리
