+++ src/basic/hooks/useCartService.ts
@@ -3,7 +3,7 @@ import { useCallback } from 'react';
 import { useCartStore } from './useCartStore';
 import type { NotificationVariant } from '../../types';
 import type { ProductWithUI } from '../constants';
-import { getRemainingStock } from '../models/entity';
+import { getRemainingStock } from '../models/product';
 
 interface UseCartServiceProps {
   products: ProductWithUI[];
